Move context-generic logging components to tracing-components crate
