+++ crates/cosmos/cosmos-client-components/src/impls/mod.rs
@@ -1,7 +1,6 @@
 pub mod channel;
 pub mod client;
 pub mod connection;
-pub mod logger;
 pub mod packet;
 pub mod queries;
 pub mod transaction;
