+++ crates/cosmos/cosmos-relayer/Cargo.toml
@@ -31,6 +31,7 @@ hermes-cosmos-test-components       = { workspace = true }
 hermes-protobuf-components          = { workspace = true }
 hermes-cli-components               = { workspace = true }
 hermes-wasm-client-components       = { workspace = true }
+hermes-tracing-components           = { workspace = true }
 tendermint                          = { workspace = true, features = ["secp256k1"] }
 tendermint-proto                    = { workspace = true }
 tendermint-rpc                      = { workspace = true, features = ["http-client", "websocket-client"] }
