+++ crates/logging/tracing-components/src/contexts/logger.rs
@@ -1,0 +1,1 @@
+pub struct TracingLogger;
