+++ crates/logging/tracing-components/src/impls/relay.rs
@@ -1,7 +1,6 @@
-use cgp_core::{Async, HasErrorType};
-use core::fmt::{Debug, Display};
+use cgp_core::Async;
+use core::fmt::Display;
 use hermes_relayer_components::chain::traits::types::chain_id::HasChainId;
-use hermes_relayer_components::chain::traits::types::message::HasMessageType;
 use hermes_relayer_components::log::traits::logger::Logger;
 use hermes_relayer_components::relay::impls::packet_clearers::receive_packet::LogClearPacketError;
 use hermes_relayer_components::relay::impls::packet_relayers::general::full_relay::LogRelayPacketAction;
@@ -13,85 +12,12 @@ use hermes_relayer_components::relay::impls::update_client::skip::LogSkipBuildUp
 use hermes_relayer_components::relay::impls::update_client::wait::LogWaitUpdateClientHeightStatus;
 use hermes_relayer_components::relay::traits::chains::HasRelayChains;
 use hermes_relayer_components::relay::traits::target::ChainTarget;
-use hermes_relayer_components::transaction::impls::estimate_fees_and_send_tx::LogSendMessagesWithSignerAndNonce;
-use hermes_relayer_components::transaction::impls::poll_tx_response::{
-    LogRetryQueryTxResponse, TxNoResponseError,
-};
-use hermes_relayer_components::transaction::traits::types::nonce::HasNonceType;
-use hermes_relayer_components::transaction::traits::types::signer::HasSignerType;
-use hermes_relayer_components::transaction::traits::types::tx_hash::HasTransactionHashType;
 use hermes_relayer_components_extra::batch::worker::LogBatchWorker;
-use tracing::{debug, error, info, trace};
+use tracing::{error, trace};
 
-pub struct HandleCosmosLogs;
+use crate::contexts::logger::TracingLogger;
 
-impl<Logging> Logger<Logging, ()> for HandleCosmosLogs
-where
-    Logging: Async,
-{
-    async fn log(_logging: &Logging, message: &str, _details: &()) {
-        info!("{message}");
-    }
-}
-
-impl<'a, Logging, Chain> Logger<Logging, LogSendMessagesWithSignerAndNonce<'a, Chain>>
-    for HandleCosmosLogs
-where
-    Logging: Async,
-    Chain: HasSignerType + HasNonceType + HasMessageType + HasChainId,
-    Chain::Signer: Debug,
-    Chain::Nonce: Debug,
-{
-    async fn log(
-        _logging: &Logging,
-        message: &str,
-        details: &LogSendMessagesWithSignerAndNonce<'a, Chain>,
-    ) {
-        trace!(
-            chain_id = %details.chain.chain_id(),
-            nonce = ?details.nonce,
-            signer = ?details.signer,
-            "{message}",
-        );
-    }
-}
-
-impl<'a, Logging, Chain> Logger<Logging, TxNoResponseError<'a, Chain>> for HandleCosmosLogs
-where
-    Logging: Async,
-    Chain: HasTransactionHashType + HasChainId,
-    Chain::TxHash: Display,
-{
-    async fn log(_logging: &Logging, message: &str, details: &TxNoResponseError<'a, Chain>) {
-        error!(
-            chain_id = %details.chain.chain_id(),
-            tx_hash = %details.tx_hash,
-            wait_timeout = ?details.wait_timeout,
-            elapsed = ?details.elapsed,
-            "{message}",
-        );
-    }
-}
-
-impl<'a, Logging, Chain> Logger<Logging, LogRetryQueryTxResponse<'a, Chain>> for HandleCosmosLogs
-where
-    Logging: Async,
-    Chain: HasTransactionHashType + HasChainId + HasErrorType,
-    Chain::TxHash: Display,
-    Chain::Error: Debug,
-{
-    async fn log(_logging: &Logging, message: &str, details: &LogRetryQueryTxResponse<'a, Chain>) {
-        debug!(
-            chain_id = %details.chain.chain_id(),
-            tx_hash = %details.tx_hash,
-            elapsed = ?details.elapsed,
-            error = ?details.error,
-            "{message}",
-        );
-    }
-}
-
-impl<'a, Logging, Relay> Logger<Logging, LogSkipRelayLockedPacket<'a, Relay>> for HandleCosmosLogs
+impl<'a, Logging, Relay> Logger<Logging, LogSkipRelayLockedPacket<'a, Relay>> for TracingLogger
 where
     Logging: Async,
     Relay: HasRelayChains,
@@ -109,7 +35,7 @@ where
     }
 }
 
-impl<'a, Logging, Relay> Logger<Logging, LogRelayPacketAction<'a, Relay>> for HandleCosmosLogs
+impl<'a, Logging, Relay> Logger<Logging, LogRelayPacketAction<'a, Relay>> for TracingLogger
 where
     Logging: Async,
     Relay: HasRelayChains,
@@ -128,7 +54,7 @@ where
     }
 }
 
-impl<'a, Logging, Relay> Logger<Logging, LogClearPacketError<'a, Relay>> for HandleCosmosLogs
+impl<'a, Logging, Relay> Logger<Logging, LogClearPacketError<'a, Relay>> for TracingLogger
 where
     Logging: Async,
     Relay: HasRelayChains,
@@ -148,7 +74,7 @@ where
     }
 }
 
-impl<'a, Logging, Relay> Logger<Logging, LogRelayPacketStatus<'a, Relay>> for HandleCosmosLogs
+impl<'a, Logging, Relay> Logger<Logging, LogRelayPacketStatus<'a, Relay>> for TracingLogger
 where
     Logging: Async,
     Relay: HasRelayChains,
@@ -190,7 +116,7 @@ where
 }
 
 impl<'a, Logging, Relay, Target> Logger<Logging, LogSkipBuildUpdateClientMessage<'a, Relay, Target>>
-    for HandleCosmosLogs
+    for TracingLogger
 where
     Logging: Async,
     Relay: HasRelayChains,
@@ -213,7 +139,7 @@ where
 }
 
 impl<'a, Logging, Relay, Target> Logger<Logging, LogWaitUpdateClientHeightStatus<'a, Relay, Target>>
-    for HandleCosmosLogs
+    for TracingLogger
 where
     Logging: Async,
     Relay: HasRelayChains,
@@ -256,7 +182,7 @@ where
 }
 
 impl<'a, Logging, Relay, Target> Logger<Logging, LogBatchWorker<'a, Relay, Target>>
-    for HandleCosmosLogs
+    for TracingLogger
 where
     Logging: Async,
     Relay: HasRelayChains,
