+++ crates/sovereign/sovereign-cosmos-relayer/Cargo.toml
@@ -28,5 +28,6 @@ hermes-wasm-client-components       = { workspace = true }
 hermes-cosmos-client-components     = { workspace = true }
 hermes-sovereign-test-components    = { workspace = true }
 hermes-protobuf-components          = { workspace = true }
+hermes-tracing-components           = { workspace = true }
 
 jsonrpsee       = { workspace = true, features = ["http-client"] }
